Support Influx/telegraf style tags
Telegraf's statsd protocol listener has added support for tags following
the metric name in comma-delimited name=value pairs before the : value
delimiter. Such as the following.

    key,tag1=value1,tag2=value2:0|c

This is a useful thing to support in addition to the DataDog's tags
extension at the end of the line because it allows for people using
things like Python's metrology that have a statsd backend that don't
understand DataDog's tags extension. Those people can just set their
metric name to include the tags, and it'll be blindly forwarded over to
gostatsd where it can be understood as tags.